doc/examples/MoonMarket/backend/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
bcrypt==4.1.2
beanie==1.25.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
ecdsa==0.19.0
email_validator==2.1.1
executing==2.0.1
fastapi-jwt==0.2.0
fastapi-users==13.0.0
fastapi-users-db-beanie==3.0.0
fastjsonschema==2.20.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lazy-model==0.2.0
makefun==1.15.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
motor==3.4.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure_eval==0.2.3
pwdlib==0.2.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
pyzmq==26.1.0
referencing==0.35.1
requests==2.31.0
rpds-py==0.20.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette
starlette
tinycss2==1.3.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.13.1
python-jose[cryptography]
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
redis>=5.0.1
aioredis==2.0.1  # For async Redis support
apify-client==1.6.4
vaderSentiment==3.3.2
openai
httpx==0.27.2